+++ Library/analogWrite.h
@@ -2,8 +2,8 @@
 #include <avr/io.h>
 #include <util/delay.h>
 
-void blink (short int times);
+void blink (uint8_t times);
 
 void clear_all_TC(void);
 
-short int analogWrite(short int apin, short int cycle) ;
+uint8_t analogWrite(uint8_t apin, uint8_t cycle) ;
